JetBrains: Cody: Move autocomplete-related actions into a dedicated subpackage
