+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/autocomplete/action/AutocompleteActionHandler.kt
@@ -1,4 +1,4 @@
-package com.sourcegraph.cody.autocomplete
+package com.sourcegraph.cody.autocomplete.action
 
 import com.intellij.openapi.actionSystem.DataContext
 import com.intellij.openapi.editor.Caret
@@ -6,6 +6,7 @@ import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.editor.actionSystem.EditorActionHandler
 import com.intellij.openapi.project.Project
 import com.sourcegraph.cody.agent.CodyAgent
+import com.sourcegraph.cody.autocomplete.InlayModelUtil
 import com.sourcegraph.cody.autocomplete.render.CodyAutocompleteElementRenderer
 import com.sourcegraph.cody.vscode.InlineAutocompleteItem
 import com.sourcegraph.utils.CodyEditorUtil
