+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/autocomplete/action/DisposeInlaysActionHandler.kt
@@ -1,8 +1,9 @@
-package com.sourcegraph.cody.autocomplete
+package com.sourcegraph.cody.autocomplete.action
 
 import com.intellij.openapi.actionSystem.DataContext
 import com.intellij.openapi.editor.Caret
 import com.intellij.openapi.editor.Editor
+import com.sourcegraph.cody.autocomplete.CodyAutocompleteManager
 
 class DisposeInlaysActionHandler : AutocompleteActionHandler() {
   override fun doExecute(editor: Editor, caret: Caret?, dataContext: DataContext?) =
