+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/autocomplete/action/TriggerAutocompleteAction.kt
@@ -1,4 +1,4 @@
-package com.sourcegraph.cody.autocomplete
+package com.sourcegraph.cody.autocomplete.action
 
 import com.intellij.openapi.editor.actionSystem.EditorAction
