+++ client/jetbrains/src/main/kotlin/com/sourcegraph/cody/autocomplete/action/TriggerAutocompleteActionHandler.kt
@@ -1,10 +1,11 @@
-package com.sourcegraph.cody.autocomplete
+package com.sourcegraph.cody.autocomplete.action
 
 import com.intellij.openapi.actionSystem.DataContext
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.editor.Caret
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.editor.actionSystem.EditorActionHandler
+import com.sourcegraph.cody.autocomplete.CodyAutocompleteManager
 import com.sourcegraph.cody.vscode.InlineCompletionTriggerKind
 import com.sourcegraph.utils.CodyEditorUtil
