+++ client/jetbrains/src/main/resources/META-INF/plugin.xml
@@ -150,18 +150,18 @@
 
         <!-- autocomplete -->
         <action id="cody.acceptAutocompleteAction"
-                class="com.sourcegraph.cody.autocomplete.AcceptCodyAutocompleteAction">
+                class="com.sourcegraph.cody.autocomplete.action.AcceptCodyAutocompleteAction">
             <keyboard-shortcut first-keystroke="TAB" keymap="$default"/>
             <override-text place="MainMenu" text="Accept Autocomplete Suggestion"/>
         </action>
 
-        <action id="cody.disposeInlays" class="com.sourcegraph.cody.autocomplete.DisposeInlaysAction">
+        <action id="cody.disposeInlays" class="com.sourcegraph.cody.autocomplete.action.DisposeInlaysAction">
             <keyboard-shortcut first-keystroke="ESCAPE" keymap="$default"/>
             <override-text place="MainMenu" text="Hide Completions"/>
         </action>
 
         <action id="cody.triggerAutocomplete"
-                class="com.sourcegraph.cody.autocomplete.TriggerAutocompleteAction">
+                class="com.sourcegraph.cody.autocomplete.action.TriggerAutocompleteAction">
             <keyboard-shortcut first-keystroke="alt BACK_SLASH" keymap="$default"/>
             <override-text place="MainMenu" text="Autocomplete"/>
         </action>
